requirements.txt
openai-whisper==20231117
pyaudio==0.2.11
pydub==0.25.1
numpy==1.24.3
scipy==1.10.1
pyttsx3==2.90
playsound==1.3.0
keyboard==0.13.5